修复 #184 中提到的问题
